pyproject.toml
[tool.poetry]
name = "custom_playbooks"
version = "0.0.1"
description = ""
authors = ["Muhammad Ali Khan <ali.khan@xgrid.co>"]

[tool.poetry.dependencies]
python = "^3.7.1"
robusta-cli = "^0.8.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
